+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - Unreleased
+
+- Fixed bug in experiment automatic seeding.
+- Migrated experiments to submodule `polypart.experiments`.
+- Fixed issue with "inf" moment calculation in partition tree statistics.
+- pyproject.toml dependency bug fix.
+
 ## [0.2.0] - 2025-12-07
 
 - Implemented IncEnu algorithm restricted to polyhedral regions.